+++ docs/changelog.md
@@ -10,9 +10,21 @@ pageflow_next_text: Contributing
 
 ## develop branch
 
-Nothing yet.
+## v1.2016052501
+
+Released Wed 25th May 2016.
+
+### Refactor
+
+* Move the code caller functionality into `ganbarodigital/php-the-missing-bits`. This is the right place for the functionality to live.
+  - `FilterBacktrace` now extends code provided by `php-the-missing-bits`
+  - `CodeCaller` now extends code provided by `php-the-missing-bits`
+  - `FilterBacktrace`, `FilterBacktraceForTwoCodeCallers` and `FilterCodeCaller` no longer support filtering partial namespaces
+  - `FilterCodeCaller::$DEFAULT_PARTIALS` is gone
 
 ## v1.2016052401
+
+Released Tue 24th May 2016.
 
 ### Fixes